requirements.txt
# Backend Dependencies
fastapi==0.135.3
uvicorn[standard]==0.44.0
websockets==16.0
python-multipart==0.0.24
python-dotenv==1.2.2

# LangChain
langchain==1.2.15
langchain-community==0.4.1
langchain-openai==1.1.12
langchain-qdrant==1.1.0
langchain-groq==1.1.2
langchain-ollama==1.0.1

# Vector Database
qdrant-client==1.17.1

# Observability
langfuse==4.0.6

# Document Processing
pypdf==6.9.2
python-docx==1.2.0
openpyxl==3.1.5
pandas==3.0.2
tiktoken==0.12.0

# Embeddings
fastembed==0.8.0

# LLM Providers
ollama==0.6.1

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
python-multipart==0.0.24

# Email
fastapi-mail==1.4.2
aiosmtplib==3.0.2
email-validator==2.2.0

# Other utilities
httpx==0.28.1
pydantic==2.12.5
pydantic-settings==2.13.1
